Refactor: Move SurveyAdapter to survey package
Moved the SurveyAdapter used in BellDashboardFragment from the `dashboard` package to the more appropriate `survey` package.

To avoid a name collision with the existing `SurveyAdapter` in the `survey` package, the moved adapter has been renamed to `SurveyTitlesAdapter`, as its purpose is to display a list of survey titles in a dialog.

The package declaration and the import statement in `BellDashboardFragment` have been updated accordingly.
